@@ -1,17 +1,18 @@
 name: youtube_player_iframe_example
 description: Demonstrates how to use the youtube_player_iframe plugin.
-publish_to: 'none'
-version: 1.1.0
+publish_to: none
+version: 4.0.3
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
-  youtube_player_iframe:
-    path: ../
+  flutter_web_plugins:
+      sdk: flutter
+  youtube_player_iframe: ^5.1.4
+  go_router: ^14.2.3
 
 dev_dependencies:
   flutter_test:
