@@ -1,20 +1,40 @@
 name: youtube_player_iframe
 description: Flutter port of the official YouTube iFrame player API. Supports web & mobile platforms.
-version: 2.3.0
+version: 5.1.5
 repository: https://github.com/sarbagyastha/youtube_player_flutter
 homepage: https://github.com/sarbagyastha/youtube_player_flutter/tree/master/packages/youtube_player_iframe
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
-  flutter: '>=3.0.0'
+  sdk: ^3.3.0
+  flutter: '>=3.19.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
-  flutter_inappwebview: ^5.4.3+7
-  url_launcher: ^6.1.2
+  meta: '>=1.7.0 <2.0.0'
+  webview_flutter: ^4.8.0
+  webview_flutter_android: ^3.16.6
+  webview_flutter_wkwebview: ^3.14.0
+  url_launcher: ^6.3.0
+  youtube_player_iframe_web: ^3.0.3
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  flutter_lints: ^4.0.0
+
+flutter:
+  assets:
+    - assets/player.html
+
+  plugin:
+    platforms:
+      web:
+        default_package: youtube_player_iframe_web
+
+platforms:
+  android:
+  ios:
+  web:
+
+screenshots:
+  - description: Youtube Player iFrame Logo
+    path: screenshots/logo.png
