@@ -2,192 +2,32 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:developer';
-
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:youtube_player_iframe/youtube_player_iframe.dart';
-
-import 'widgets/meta_data_section.dart';
-import 'widgets/play_pause_button_bar.dart';
-import 'widgets/player_state_section.dart';
-import 'widgets/source_input_section.dart';
-import 'widgets/volume_slider.dart';
+import 'package:youtube_player_iframe_example/router.dart';
+import 'package:flutter_web_plugins/url_strategy.dart';
 
 Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  runApp(YoutubeApp());
+  usePathUrlStrategy();
+  runApp(const YoutubeApp());
 }
 
 ///
 class YoutubeApp extends StatelessWidget {
+  const YoutubeApp({super.key});
+
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    final colorScheme = ColorScheme.fromSeed(
+      seedColor: Colors.green,
+      dynamicSchemeVariant: DynamicSchemeVariant.expressive,
+      brightness: Brightness.dark,
+    );
+
+    return MaterialApp.router(
       title: 'Youtube Player IFrame Demo',
-      theme: ThemeData(
-        brightness: Brightness.dark,
-        primarySwatch: Colors.deepPurple,
-        scaffoldBackgroundColor: Colors.black,
-      ),
+      theme: ThemeData.from(colorScheme: colorScheme),
       debugShowCheckedModeBanner: false,
-      home: YoutubeAppDemo(),
+      routerConfig: router,
     );
   }
-}
-
-///
-class YoutubeAppDemo extends StatefulWidget {
-  @override
-  _YoutubeAppDemoState createState() => _YoutubeAppDemoState();
-}
-
-class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
-  late YoutubePlayerController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = YoutubePlayerController(
-      initialVideoId: 'tcodrIK2P_I',
-      params: const YoutubePlayerParams(
-        playlist: [
-          'nPt8bK2gbaU',
-          'K18cpp_-gP8',
-          'iLnmTe5Q2Qw',
-          '_WoCV4c6XOE',
-          'KmzdUe0RSJo',
-          '6jZDSSZZxjQ',
-          'p2lYr3vM_1w',
-          '7QUtEmBT_-w',
-          '34_PXCzGw1M',
-        ],
-        startAt: const Duration(minutes: 1, seconds: 36),
-        showControls: true,
-        showFullscreenButton: true,
-        desktopMode: false,
-        privacyEnhanced: true,
-        useHybridComposition: true,
-      ),
-    );
-    _controller.onEnterFullscreen = () {
-      SystemChrome.setPreferredOrientations([
-        DeviceOrientation.landscapeLeft,
-        DeviceOrientation.landscapeRight,
-      ]);
-      log('Entered Fullscreen');
-    };
-    _controller.onExitFullscreen = () {
-      log('Exited Fullscreen');
-    };
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    const player = YoutubePlayerIFrame();
-    return YoutubePlayerControllerProvider(
-      // Passing controller to widgets below.
-      controller: _controller,
-      child: Scaffold(
-        body: LayoutBuilder(
-          builder: (context, constraints) {
-            if (kIsWeb && constraints.maxWidth > 800) {
-              return Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  const Expanded(child: player),
-                  const SizedBox(
-                    width: 500,
-                    child: SingleChildScrollView(
-                      child: Controls(),
-                    ),
-                  ),
-                ],
-              );
-            }
-            return ListView(
-              children: [
-                Stack(
-                  children: [
-                    player,
-                    Positioned.fill(
-                      child: YoutubeValueBuilder(
-                        controller: _controller,
-                        builder: (context, value) {
-                          return AnimatedCrossFade(
-                            firstChild: const SizedBox.shrink(),
-                            secondChild: Material(
-                              child: DecoratedBox(
-                                decoration: BoxDecoration(
-                                  image: DecorationImage(
-                                    image: NetworkImage(
-                                      YoutubePlayerController.getThumbnail(
-                                        videoId:
-                                            _controller.params.playlist.first,
-                                        quality: ThumbnailQuality.medium,
-                                      ),
-                                    ),
-                                    fit: BoxFit.fitWidth,
-                                  ),
-                                ),
-                                child: const Center(
-                                  child: CircularProgressIndicator(),
-                                ),
-                              ),
-                            ),
-                            crossFadeState: value.isReady
-                                ? CrossFadeState.showFirst
-                                : CrossFadeState.showSecond,
-                            duration: const Duration(milliseconds: 300),
-                          );
-                        },
-                      ),
-                    ),
-                  ],
-                ),
-                const Controls(),
-              ],
-            );
-          },
-        ),
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    _controller.close();
-    super.dispose();
-  }
-}
-
-///
-class Controls extends StatelessWidget {
-  ///
-  const Controls();
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          _space,
-          MetaDataSection(),
-          _space,
-          SourceInputSection(),
-          _space,
-          PlayPauseButtonBar(),
-          _space,
-          VolumeSlider(),
-          _space,
-          PlayerStateSection(),
-        ],
-      ),
-    );
-  }
-
-  Widget get _space => const SizedBox(height: 10);
 }
