@@ -9,20 +9,21 @@ import '../utils/youtube_player_controller.dart';
 
 /// A widget to display playback speed changing button.
 class PlaybackSpeedButton extends StatefulWidget {
+  /// Creates [PlaybackSpeedButton] widget.
+  const PlaybackSpeedButton({
+    super.key,
+    this.controller,
+    this.icon,
+  });
+
   /// Overrides the default [YoutubePlayerController].
   final YoutubePlayerController? controller;
 
   /// Defines icon for the button.
   final Widget? icon;
 
-  /// Creates [PlaybackSpeedButton] widget.
-  const PlaybackSpeedButton({
-    this.controller,
-    this.icon,
-  });
-
   @override
-  _PlaybackSpeedButtonState createState() => _PlaybackSpeedButtonState();
+  State<PlaybackSpeedButton> createState() => _PlaybackSpeedButtonState();
 }
 
 class _PlaybackSpeedButtonState extends State<PlaybackSpeedButton> {
@@ -48,17 +49,6 @@ class _PlaybackSpeedButtonState extends State<PlaybackSpeedButton> {
   Widget build(BuildContext context) {
     return PopupMenuButton<double>(
       onSelected: _controller.setPlaybackRate,
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(8.0, 8.0, 0.0, 8.0),
-        child: widget.icon ??
-            Image.asset(
-              'assets/speedometer.webp',
-              package: 'youtube_player_flutter',
-              width: 20.0,
-              height: 20.0,
-              color: Colors.white,
-            ),
-      ),
       tooltip: 'PlayBack Rate',
       itemBuilder: (context) => [
         _popUpItem('2.0x', PlaybackRate.twice),
@@ -70,14 +60,25 @@ class _PlaybackSpeedButtonState extends State<PlaybackSpeedButton> {
         _popUpItem('0.5x', PlaybackRate.half),
         _popUpItem('0.25x', PlaybackRate.quarter),
       ],
+      child: Padding(
+        padding: const EdgeInsets.fromLTRB(8, 8, 0, 8),
+        child: widget.icon ??
+            Image.asset(
+              'assets/speedometer.webp',
+              package: 'youtube_player_flutter',
+              width: 20.0,
+              height: 20.0,
+              color: Colors.white,
+            ),
+      ),
     );
   }
 
   PopupMenuEntry<double> _popUpItem(String text, double rate) {
     return CheckedPopupMenuItem(
       checked: _controller.value.playbackRate == rate,
-      child: Text(text),
       value: rate,
+      child: Text(text),
     );
   }
 }
