@@ -1,5 +1,4 @@
 import 'dart:io';
-import 'dart:ui' show window;
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -59,21 +58,22 @@ void main() {
 
   testWidgets('Create Youtube Player', (WidgetTester tester) async {
     provideMockedNetworkImages(() async {
-      var _controller = createController();
+      final controller = createController();
 
-      await tester.pumpWidget(buildPlayer(controller: _controller));
+      await tester.pumpWidget(buildPlayer(controller: controller));
     });
   });
 }
 
 class TestApp extends StatelessWidget {
-  final Widget child;
-  final TextDirection textDirection;
-
-  TestApp({
+  const TestApp({
+    super.key,
     this.textDirection = TextDirection.ltr,
     required this.child,
   });
+
+  final Widget child;
+  final TextDirection textDirection;
 
   @override
   Widget build(BuildContext context) {
@@ -84,7 +84,7 @@ class TestApp extends StatelessWidget {
         DefaultMaterialLocalizations.delegate,
       ],
       child: MediaQuery(
-        data: MediaQueryData.fromWindow(window),
+        data: MediaQueryData.fromView(View.of(context)),
         child: Directionality(
           textDirection: textDirection,
           child: child,
@@ -94,7 +94,7 @@ class TestApp extends StatelessWidget {
   }
 }
 
-R provideMockedNetworkImages<R>(R body()) {
+R provideMockedNetworkImages<R>(R Function() body) {
   return HttpOverrides.runZoned(
     body,
     createHttpClient: (_) => _createMockImageHttpClient(_, _transparentImage),
