@@ -7,10 +7,15 @@ import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
 ///
 class MetaDataSection extends StatelessWidget {
+  const MetaDataSection({super.key});
+
   @override
   Widget build(BuildContext context) {
     return YoutubeValueBuilder(
-      buildWhen: (o, n) => o.metaData != n.metaData,
+      buildWhen: (o, n) {
+        return o.metaData != n.metaData ||
+            o.playbackQuality != n.playbackQuality;
+      },
       builder: (context, value) {
         return Column(
           crossAxisAlignment: CrossAxisAlignment.start,
@@ -41,13 +46,13 @@ class MetaDataSection extends StatelessWidget {
                       items: PlaybackRate.all
                           .map(
                             (rate) => DropdownMenuItem(
+                              value: rate,
                               child: Text(
                                 '${rate}x',
                                 style: const TextStyle(
                                   fontWeight: FontWeight.w300,
                                 ),
                               ),
-                              value: rate,
                             ),
                           )
                           .toList(),
@@ -76,18 +81,17 @@ class _Text extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return RichText(
-      text: TextSpan(
+    return Text.rich(
+      TextSpan(
         text: '$title : ',
-        style: const TextStyle(
-          fontWeight: FontWeight.bold,
-        ),
+        style: Theme.of(context).textTheme.labelLarge,
         children: [
           TextSpan(
             text: value,
-            style: const TextStyle(
-              fontWeight: FontWeight.w300,
-            ),
+            style: Theme.of(context)
+                .textTheme
+                .labelMedium!
+                .copyWith(fontWeight: FontWeight.w300),
           ),
         ],
       ),
