@@ -7,7 +7,10 @@ import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
 ///
 class PlayPauseButtonBar extends StatelessWidget {
+  PlayPauseButtonBar({super.key});
+
   final ValueNotifier<bool> _isMuted = ValueNotifier(false);
+
   @override
   Widget build(BuildContext context) {
     return Row(
@@ -25,13 +28,11 @@ class PlayPauseButtonBar extends StatelessWidget {
                     ? Icons.pause
                     : Icons.play_arrow,
               ),
-              onPressed: value.isReady
-                  ? () {
-                      value.playerState == PlayerState.playing
-                          ? context.ytController.pause()
-                          : context.ytController.play();
-                    }
-                  : null,
+              onPressed: () {
+                value.playerState == PlayerState.playing
+                    ? context.ytController.pauseVideo()
+                    : context.ytController.playVideo();
+              },
             );
           },
         ),
