@@ -7,7 +7,7 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 
-import '../controller.dart';
+import '../controller/youtube_player_controller.dart';
 import '../player_value.dart';
 import 'youtube_value_provider.dart';
 
@@ -35,14 +35,14 @@ class YoutubeValueBuilder extends StatefulWidget {
   ///
   /// The [builder] must not be null.
   const YoutubeValueBuilder({
-    Key? key,
+    super.key,
     required this.builder,
     this.buildWhen,
     this.controller,
-  }) : super(key: key);
+  });
 
   @override
-  _YoutubeValueBuilderState createState() => _YoutubeValueBuilderState();
+  State<YoutubeValueBuilder> createState() => _YoutubeValueBuilderState();
 }
 
 class _YoutubeValueBuilderState extends State<YoutubeValueBuilder> {
@@ -92,8 +92,9 @@ class _YoutubeValueBuilderState extends State<YoutubeValueBuilder> {
     _subscription = _controller!.listen(
       (value) {
         if (widget.buildWhen?.call(_previousValue, value) ?? true) {
+          if (!mounted) return;
           _child = widget.builder(context, value);
-          if (mounted) setState(() {});
+          setState(() {});
         }
         _previousValue = value;
       },
